requirements.txt
Django==5.2.1
djangorestframework==3.16.0
pillow==11.2.1
django-extensions==4.1
django-polymorphic==4.1.0
Faker==37.3.0
django-cte==1.3.3
django-rest-polymorphic==0.1.10
